Strip starter UI to a single "Let's build." root screen (#1)
<!-- CURSOR_AGENT_PR_BODY_BEGIN -->
## Summary

This change removes the Expo Router tab scaffold and demo screens. The app now has a single root route (`app/index.tsx`) that shows **Let's build.** centered on the screen.

## Changes

- Replaced `(tabs)` layout with a minimal stack: `index` and `+not-found`.
- Removed unused demo components (parallax home, explore tab, collapsibles, tab bar helpers, icon symbol wrappers, etc.).
- Dropped `SpaceMono` font loading from the root layout since nothing uses it anymore.

`ThemedText` / `ThemedView` remain so light/dark theming still applies.
<!-- CURSOR_AGENT_PR_BODY_END -->

<div><a href="https://cursor.com/agents/bc-3081481f-7509-4a7c-b221-cd8107574d97"><picture><source media="(prefers-color-scheme: dark)" srcset="https://cursor.com/assets/images/open-in-web-dark.png"><source media="(prefers-color-scheme: light)" srcset="https://cursor.com/assets/images/open-in-web-light.png"><img alt="Open in Web" width="114" height="28" src="https://cursor.com/assets/images/open-in-web-dark.png"></picture></a>&nbsp;<a href="https://cursor.com/background-agent?bcId=bc-3081481f-7509-4a7c-b221-cd8107574d97"><picture><source media="(prefers-color-scheme: dark)" srcset="https://cursor.com/assets/images/open-in-cursor-dark.png"><source media="(prefers-color-scheme: light)" srcset="https://cursor.com/assets/images/open-in-cursor-light.png"><img alt="Open in Cursor" width="131" height="28" src="https://cursor.com/assets/images/open-in-cursor-dark.png"></picture></a>&nbsp;</div>
