+++ app/index.tsx
@@ -1,0 +1,20 @@
+import { StyleSheet } from 'react-native';
+
+import { ThemedText } from '@/components/ThemedText';
+import { ThemedView } from '@/components/ThemedView';
+
+export default function Index() {
+  return (
+    <ThemedView style={styles.container}>
+      <ThemedText type="title">Let's build.</ThemedText>
+    </ThemedView>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
